apphub-service/apphub-service-skyxplore/apphub-service-skyxplore-data/src/main/java/com/github/saphyra/apphub/service/skyxplore/data/save_game/dao/building_allocation/BuildingAllocationDao.java
package com.github.saphyra.apphub.service.skyxplore.data.save_game.dao.building_allocation;

import com.github.saphyra.apphub.api.skyxplore.model.game.BuildingAllocationModel;
import com.github.saphyra.apphub.lib.common_util.AbstractDao;
import com.github.saphyra.apphub.lib.common_util.converter.UuidConverter;
import org.springframework.stereotype.Component;

import java.util.UUID;

@Component
public class BuildingAllocationDao extends AbstractDao<BuildingAllocationEntity, BuildingAllocationModel, String, BuildingAllocationRepository> {
    private final UuidConverter uuidConverter;

    public BuildingAllocationDao(BuildingAllocationConverter converter, BuildingAllocationRepository repository, UuidConverter uuidConverter) {
        super(converter, repository);
        this.uuidConverter = uuidConverter;
    }

    public void deleteByGameId(UUID gameId) {
        repository.deleteByGameId(uuidConverter.convertDomain(gameId));
    }

    public void deleteById(UUID buildingAllocationId) {
        deleteById(uuidConverter.convertDomain(buildingAllocationId));
    }
}
